requirements.txt
Django~=4.1.1
djangorestframework~=3.14.0
python-dotenv~=0.21.0
pycodestyle~=2.9.1
pylint~=2.15.3
psycopg2~=2.9.3
Pillow~=9.2.0
djangorestframework-simplejwt~=5.2.1
djoser~=2.1.0
